# paper/artifacts/phase2/swagger_openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SupplyTwinAI API",
    "description": "Autonomous Multi-Agent Supply Chain Digital Twin REST API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_v1_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "List Orders",
        "operationId": "list_orders_api_v1_orders__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponse"
                  },
                  "title": "Response List Orders Api V1 Orders  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get Order",
        "operationId": "get_order_api_v1_orders__order_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shipments/": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "summary": "List Shipments",
        "operationId": "list_shipments_api_v1_shipments__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShipmentResponse"
                  },
                  "title": "Response List Shipments Api V1 Shipments  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shipments/{shipment_id}": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "summary": "Get Shipment",
        "operationId": "get_shipment_api_v1_shipments__shipment_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shipment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/etl/load": {
      "post": {
        "tags": [
          "ETL Management (Admin)"
        ],
        "summary": "Trigger Etl Load",
        "operationId": "trigger_etl_load_api_v1_etl_load_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "v3_etl_dir",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "datasets/v3_etl",
              "title": "V3 Etl Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ETLLoadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/etl/history": {
      "get": {
        "tags": [
          "ETL Management (Admin)"
        ],
        "summary": "Get Etl History",
        "operationId": "get_etl_history_api_v1_etl_history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ETLRunResponse"
                  },
                  "title": "Response Get Etl History Api V1 Etl History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/config": {
      "get": {
        "tags": [
          "Admin Configuration"
        ],
        "summary": "Get Agent Config",
        "operationId": "get_agent_config_api_v1_admin_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConfigResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Admin Configuration"
        ],
        "summary": "Update Agent Config",
        "operationId": "update_agent_config_api_v1_admin_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentConfigResponse": {
        "properties": {
          "config_id": {
            "type": "integer",
            "title": "Config Id"
          },
          "risk_weight_supplier": {
            "type": "number",
            "title": "Risk Weight Supplier"
          },
          "risk_weight_shipment": {
            "type": "number",
            "title": "Risk Weight Shipment"
          },
          "risk_weight_inventory": {
            "type": "number",
            "title": "Risk Weight Inventory"
          },
          "risk_weight_external": {
            "type": "number",
            "title": "Risk Weight External"
          },
          "threshold_low_medium": {
            "type": "number",
            "title": "Threshold Low Medium"
          },
          "threshold_medium_high": {
            "type": "number",
            "title": "Threshold Medium High"
          },
          "alert_threshold_recommendation": {
            "type": "number",
            "title": "Alert Threshold Recommendation"
          },
          "updated_by": {
            "type": "string",
            "title": "Updated By"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "config_id",
          "risk_weight_supplier",
          "risk_weight_shipment",
          "risk_weight_inventory",
          "risk_weight_external",
          "threshold_low_medium",
          "threshold_medium_high",
          "alert_threshold_recommendation",
          "updated_by",
          "updated_at"
        ],
        "title": "AgentConfigResponse"
      },
      "AgentConfigUpdate": {
        "properties": {
          "risk_weight_supplier": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk Weight Supplier"
          },
          "risk_weight_shipment": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk Weight Shipment"
          },
          "risk_weight_inventory": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk Weight Inventory"
          },
          "risk_weight_external": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk Weight External"
          },
          "threshold_low_medium": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Threshold Low Medium"
          },
          "threshold_medium_high": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Threshold Medium High"
          },
          "alert_threshold_recommendation": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Alert Threshold Recommendation"
          }
        },
        "type": "object",
        "required": [
          "risk_weight_supplier",
          "risk_weight_shipment",
          "risk_weight_inventory",
          "risk_weight_external",
          "threshold_low_medium",
          "threshold_medium_high",
          "alert_threshold_recommendation"
        ],
        "title": "AgentConfigUpdate"
      },
      "Body_login_api_v1_auth_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_api_v1_auth_login_post"
      },
      "ETLLoadResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          }
        },
        "type": "object",
        "required": [
          "status",
          "duration_seconds",
          "seed",
          "counts"
        ],
        "title": "ETLLoadResponse"
      },
      "ETLRunResponse": {
        "properties": {
          "run_id": {
            "type": "integer",
            "title": "Run Id"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "raw_file_name": {
            "type": "string",
            "title": "Raw File Name"
          },
          "loaded_records": {
            "type": "integer",
            "title": "Loaded Records"
          },
          "rejected_records": {
            "type": "integer",
            "title": "Rejected Records"
          },
          "flagged_records": {
            "type": "integer",
            "title": "Flagged Records"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "seed",
          "raw_file_name",
          "loaded_records",
          "rejected_records",
          "flagged_records",
          "duration_seconds",
          "status",
          "created_at"
        ],
        "title": "ETLRunResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderResponse": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "global_id": {
            "type": "string",
            "title": "Global Id"
          },
          "customer_id": {
            "type": "integer",
            "title": "Customer Id"
          },
          "customer_global_id": {
            "type": "string",
            "title": "Customer Global Id"
          },
          "order_date": {
            "type": "string",
            "format": "date-time",
            "title": "Order Date"
          },
          "order_status": {
            "type": "string",
            "title": "Order Status"
          },
          "market": {
            "type": "string",
            "title": "Market"
          },
          "order_region": {
            "type": "string",
            "title": "Order Region"
          },
          "destination_country": {
            "type": "string",
            "title": "Destination Country"
          },
          "destination_city": {
            "type": "string",
            "title": "Destination City"
          },
          "destination_state": {
            "type": "string",
            "title": "Destination State"
          },
          "customer_latitude": {
            "type": "number",
            "title": "Customer Latitude"
          },
          "customer_longitude": {
            "type": "number",
            "title": "Customer Longitude"
          },
          "assigned_warehouse_id": {
            "type": "integer",
            "title": "Assigned Warehouse Id"
          },
          "assigned_warehouse_id_is_synthetic": {
            "type": "boolean",
            "title": "Assigned Warehouse Id Is Synthetic"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "global_id",
          "customer_id",
          "customer_global_id",
          "order_date",
          "order_status",
          "market",
          "order_region",
          "destination_country",
          "destination_city",
          "destination_state",
          "customer_latitude",
          "customer_longitude",
          "assigned_warehouse_id",
          "assigned_warehouse_id_is_synthetic",
          "is_simulated"
        ],
        "title": "OrderResponse"
      },
      "ShipmentResponse": {
        "properties": {
          "shipment_id": {
            "type": "integer",
            "title": "Shipment Id"
          },
          "global_id": {
            "type": "string",
            "title": "Global Id"
          },
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "order_global_id": {
            "type": "string",
            "title": "Order Global Id"
          },
          "shipping_mode": {
            "type": "string",
            "title": "Shipping Mode"
          },
          "days_scheduled": {
            "type": "integer",
            "title": "Days Scheduled"
          },
          "days_real": {
            "type": "integer",
            "title": "Days Real"
          },
          "delivery_status": {
            "type": "string",
            "title": "Delivery Status"
          },
          "shipping_date": {
            "type": "string",
            "format": "date-time",
            "title": "Shipping Date"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated"
          }
        },
        "type": "object",
        "required": [
          "shipment_id",
          "global_id",
          "order_id",
          "order_global_id",
          "shipping_mode",
          "days_scheduled",
          "days_real",
          "delivery_status",
          "shipping_date",
          "is_simulated"
        ],
        "title": "ShipmentResponse"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "role"
        ],
        "title": "TokenResponse"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}